Major bug(somewhere!!)
Everything broke, server not running. Tried to uninstall the express and
@types/express packages and install matching ones, but it did not work.
Version @5.0.1 to be exact

Other changes include:
- created a changelog file, which is basically a dump of all my git
  commits
- created a tree.txt file, which shows my repo files structure
- restructured logic of the users controllers, now having one for auth
  login and another for details controll.
