+++ backend/package.json
@@ -23,28 +23,23 @@
     "@types/cors": "^2.8.17",
     "@types/dotenv": "^6.1.1",
     "@types/express": "^5.0.1",
-    "@types/jest": "^29.5.14",
     "@types/joi": "^17.2.2",
     "@types/jsonwebtoken": "^9.0.9",
     "@types/nodemailer": "^6.4.17",
     "@types/nodemon": "^1.19.6",
     "@types/passport": "^1.0.17",
-    "@types/supertest": "^6.0.3",
     "@types/uuid": "^10.0.0",
     "bcrypt": "^5.1.1",
     "concurrently": "^9.1.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
-    "express": "^5.1.0",
-    "jest": "^29.7.0",
+    "express": "^5.0.1",
     "joi": "^17.13.3",
     "jsonwebtoken": "^9.0.2",
     "mysql2": "^3.12.0",
     "nodemailer": "^6.10.0",
     "nodemon": "^3.1.9",
     "passport": "^0.7.0",
-    "supertest": "^7.1.0",
-    "ts-jest": "^29.2.6",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
