+++ backend/src/api-v1/models/users.models.ts
@@ -1,0 +1,22 @@
+
+export interface Users {
+    id:number,
+    username:string,
+    email:string,
+    phoneNumber:string,
+    password:string,
+    createdAt:string,
+    isEmailVerified:string,
+    isDeleetd:string
+}
+
+export interface UserDetails {
+    id:number,
+    userId:number,
+    gender:string,
+    dob:string,
+    profilePic:string,
+    lastUpdated:string
+}
+
+
