+++ backend/src/api-v1/validators/users.validators.ts
@@ -135,29 +135,73 @@ export const updateUserSchema = Joi.object({
         'string.min':'Username should have a minimum length of {#length} characters',
         'string.max':'Username shoulf have a maximum length of {#length} characters'
     }),
-email:Joi.string()
-    .email({
-        minDomainSegments:2,
-        tlds:{
-            allow:['com','net','ke']
-        }
-    })
+    email:Joi.string()
+        .email({
+            minDomainSegments:2,
+            tlds:{
+                allow:['com','net','ke']
+            }
+        })
+        .messages({
+            'string.base':'Email must be a string',
+            'string.email':'Email can only have two domains, e.g example.com whose tlds can either be ".com", ".net" or ".ke"'
+        }),
+    phoneNumber:Joi.string()
+        .max(10)
+        .messages({
+            'string.base':'Phone number must be a string',
+            'string.max':'Phone number must have a maximum of {#length} digits'
+        }),
+    password:Joi.string()
+        .pattern(
+            new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{6,10}$')
+        )
+        .messages({
+            'string.base':'Password should be of type text',
+            'string.pattern.base':'Password must contain at least one lowercase letter, one uppercase letter, one digit, and one special character'
+        }),
+    gender:Joi.string()
+    .required()
     .messages({
-        'string.base':'Email must be a string',
-        'string.email':'Email can only have two domains, e.g example.com whose tlds can either be ".com", ".net" or ".ke"'
+        'string.base':'Gender should be of type text',
+        'any.required':'Gender is required, kindly choose between male,female or other.'
     }),
-phoneNumber:Joi.string()
-    .max(10)
+    dob:Joi.date()
+        .min('1-1-1738')
+        .max('now')
+        .messages({
+            'any.required':'A valid date is required',
+            'date.min':'Date cannot be prior to 1-1-1738',
+            'date.max':'Date cannot be later than {#now}'
+        }),
+    profilePic:Joi.string()
+        .required()
+        .message(
+            'A profile picture is required',
+        )
+
+})
+
+
+export const userDetailsSchema = Joi.object({
+    gender:Joi.string()
+    .required()
     .messages({
-        'string.base':'Phone number must be a string',
-        'string.max':'Phone number must have a maximum of {#length} digits'
+        'string.base':'Gender should be of type text',
+        'any.required':'Gender is required, kindly choose between male,female or other.'
     }),
-password:Joi.string()
-    .pattern(
-        new RegExp('^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{6,10}$')
-    )
-    .messages({
-        'string.base':'Password should be of type text',
-        'string.pattern.base':'Password must contain at least one lowercase letter, one uppercase letter, one digit, and one special character'
-    })
+    dob:Joi.date()
+        .min('1-1-1738')
+        .max('now')
+        .messages({
+            'any.required':'A valid date is required',
+            'date.min':'Date cannot be prior to 1-1-1738',
+            'date.max':'Date cannot be later than {#now}'
+        }),
+    profilePic:Joi.string()
+        .required()
+        .message(
+            'A profile picture is required'
+        )
+
 })
